This page contains the documentation for various useful functions in the NMRInversions package.

Tip

From the Julia command line, you can enter '?', followed by the name of any function, struct, or object you want to learn more about (try it!). After typing using NMRInversions in the Julia console, this feature will work for all the functions mentioned below.

Info

In Julia, function definitions look like this:

foo(x, y, z ; a, b)

For the example above, foo is the name of the function, and the contents of the parenteses are the arguments. Within the parenteses, we got two types of arguments:

Positional arguments

Everything that appears before the semicolon ; (x,y and z in this example) is necessary, and must be given in a specific order.

Keyword arguments.

Everything that appears after the semicolon ; (b and a in this example) is optional, and can be given in any order, but its name must be specified.

In the example above, we can call the function by typing foo(1, 2, 3), (in which case, x=1, y=2, z=3, and the default value for a and b will be used). You can also call the function by typing foo(1, 2, 3, a=3), to specify the value for a, or by typing foo(1, 2, 3, b=3, a = 2), to specify the value for both a and b.

Sometimes, if there are many keyword arguments, we write the function as foo(x, y ; kwargs...). For the necessary arguments before the semicolon, order matters. For the keyword arguments after the semicolon, order does not matter, but the name of each argument must be specified. For more information, please refer to this link.

Importing data

This package offers some functions to import NMR experiment data of various formats. Alternatively, you can of course import your data however you see fit. If a format you're working with is not yet supported, please submit an issue and we'll work on it.

The most basic use case would be using data saved in a csv format, where there are only two columns, one for your x-axis (time for relaxation and b-factor for diffusion) and one for your y-axis (signal intensity).

NMRInversions.import_csv — Function

import_csv(seq, file)

Import data from a CSV file.

The 1st column should include the x data points (time or b-factor). The 2nd column should be the y data points. For imaginary numbers, the 2nd and 3rd columns can be used for the real and imaginary part.

The function reads the file and returns an ExperimentData structure.

seq is the 1D pulse sequence (e.g. IR, CPMG, PFG)

file is the path to the CSV file which contains the data (x, y) in two respective columns.

The function can be called without the seq argument, and the output will be the x and y vectors (use it like, x,y =import_csv()). Alternatively, the function can also be called with only the seq argument, in which case a file dialog will open to select the file (use it like, data = import_csv(IR))

Please note that this function will just import the data as is, without any unit conversions. Ensure that your x-axis is in seconds for relaxation, or in seconds/metres² * 1e-9 for PFG experiments.

If you're using a spinsolve instrument, you can use the import_spinsolve function. This one requires two files as an input. The aqcu.par is automatically exported by SpinsolveExpert, but you might have to export your data file manually in a csv format.

NMRInversions.import_spinsolve — Method

import_spinsolve(files)

Import data from a Spinsolve expert experiment. Two paths must be provided as follows (order is not important):

files = [.../datafile.csv , .../acqu.par]

Calling this function without an argument by typing import_spinsolve() will open a file dialogue to select the files.

The function reads the acqu.par.bak file to get the acquisition parameters, and the .dat file to get the data. The function returns an ExperimentData structure.

For geospec instruments, you can export your raw data as a text file. That text file can be read by the import_geospec function.

NMRInversions.import_geospec — Function

import_geospec(dir)

Import data from a .txt format, as exported by Geospec instruments.

The function reads the relevant information, performs a phase correction on the data, and returns an ExperimentData structure.

Calling this function without an argument by typing import_geospec() will open a file dialog to select the .txt file.

For.tnt files from Tecmag consoles, import_tnt can be used. It should be noted however, that since pulse sequence implementations might vary, automatic parsing of the data into a ExperimentData structure might fail. Therefore, the more manual read_tnt_data and read_tnt_header functions can be used for fine-tuned cases. If that sounds confusing, please submit an issue and we'll work on it!

NMRInversions.import_tnt — Function

import_tnt(seq, filename)

Read data from a tecmag .tnt file, and return a ExperimentData object with all the relevant information.

This function is experimental, and since tecmag consoles work mainly with custom pulse sequences, it might not cover some cases. If it does not work for you, please submit an issue.

Necessary (positional) arguments:

seq is the pulse sequence, i.e. IR, SR, CPMG, (CPMG,IR), (CPMG,SR).

filename is the path of the file containing the data.

Optional (keyword) arguments:

echotime is the string that is used to denote the Echo time as

defined in the pulse sequence. Default value is "Echo_Time".

Calling this function without a filename argument, e.g.: import_tnt(seq), will open a file dialogue to select the .tnt file.

NMRInversions.read_tnt_data — Function

read_tnt_data(filename::String)

Read raw NMR data from a Tecmag .tnt file. Returns a ComplexF32 array. Shape is determined by npts (all 4 dims), which controls how much data was written, regardless of actual_npts.

NMRInversions.read_tnt_header — Function

read_tnt_header(filename::String)

Read TMAG parameters from a .tnt file and return them as a Dict. Covers the full 1024-byte TECMAG structure per spec.

Inversion

The most important function is invert(), which is the main function of the package. It works as follows:

NMRInversions.invert — Function

invert(input::ExperimentData{D}; kwargs...)

Receives ExperimentData, returns InversionData.

The keyword arguments are:

axes : a D-dimensional tuple where each element is an array containing the

range of the output (e.g. relaxation times or diffusion coefficients) in every corresponding dimension. nothing can be passed instead of an array in any of the tuple elements, in which case a sensible array will be generated automatically (that is also the default option).

alpha : can either be a single scalar value (e.g. 0.65) or an AlphaOptimizer.

solver : is the RegularizationSolver used to solve the optimization problem.

silent : whether the function should be printing information while it runs

(defaults to false).

scale : whether the input data should be scaled to a maximum value of 1

(you may also call that "normalizing" the data).

Can be used in exactly the same way, whether the data is 1D or 2D.

Finding alpha

Here we provide two options for finding the optimal value for alpha, namely Generalized Cross Validation (GCV) or L-curve. Generally gcv seems slightly more reliable in NMR, but it's far from perfect, so it's good to have alternatives and cross-check. The following methods can be used as inputs for the alpha argument in the invert function:

NMRInversions.GCV — Method

GCV()

Constructor for the GCV method described in Mitchell 2012. No arguments required.

NMRInversions.GCV — Method

GCV(start; kwargs...)

Constructor for finding the optimal alpha value via GCV score box optimization, given a starting value.

Necessary (positional) arguments:

start is the starting alpha value.

Choose something sensible, usually a value between 0.1 and 10 would work well.

Optional (keyword) arguments:

algorithm determines which method will be used by Optim.jl to solve the problem.

Default is LBFGS(). Only first order optimizers can be chosen here. For more details, refer to Optim.jl documentation.

opts an Optim.Options() structure which can provide some preferences to the solver.

NMRInversions.GCV — Method

GCV(lower, upper ; kwargs...)

Constructor for finding the optimal alpha value via GCV score univariate optimization, given some lower and upper bounds.

Necessary (positional) arguments:

lower is the lower bound, or lowest alpha value the algorighm will consider.

upper is the upper bound, or highest alpha value the algorighm will consider.

Optional (keyword) arguments:

algorithm determines which method will be used by Optim.jl to solve the problem.

Default is Brent() and the only alternative is GoldenSection(), which is normally slower.

abs_tol determines what's the smallest absolute change in the GCV score the algorithm should care about.

Default is 1e-3.

rel_tol determines what's the smallest relative change in the GCV score the algorithm should care about.

Default is 1e-1.

NMRInversions.LCurve — Method

LCurve(lowest_value, highest_value, number_of_steps)

Constructor for testing all L-curve curvatures between some bounds and picking the optimal.

lowest_value is the lowest alpha value.

highest_value is the highest alpha value.

number_of_steps is the number of alpha values that will be tested.

This is a very crude and rather slow method, mostly for demonstration purposes.

NMRInversions.LCurve — Method

LCurve(start; kwargs...)

Constructor for finding the optimal alpha value via LCurve curvature box optimization, given a starting value.

Necessary (positional) arguments:

start is the starting alpha value.

Choose something sensible, usually a value between 0.1 and 10 would work well.

Optional (keyword) arguments:

algorithm determines which method will be used by Optim.jl to solve the problem.

Default is LBFGS(). Only first order optimizers can be chosen here. For more details, refer to Optim.jl documentation.

opts an Optim.Options() structure which can provide some preferences to the solver.

Please have a look here.

NMRInversions.LCurve — Method

LCurve(lower, upper ; kwargs...)

Constructor for finding the optimal alpha value via LCurve curvature univariate optimization, given some lower and upper bounds.

Necessary (positional) arguments:

lower is the lower bound, or lowest alpha value the algorighm will consider.

upper is the upper bound, or highest alpha value the algorighm will consider.

Optional (keyword) arguments:

algorithm determines which method will be used by Optim.jl to solve the problem.

Default is Brent() and the only alternative is GoldenSection(), which is normally slower.

abs_tol determines what's the smallest absolute change in the LCurve score the algorithm should care about.

Default is 1e-3.

rel_tol determines what's the smallest relative change in the LCurve score the algorithm should care about.

Default is 1e-1.

The GCV() method [2] usually involves the least amount of function calls and it is thus much faster, thus used as the default option. If you want more precision, the univariate or box methods should be used instead. Note that GCV() will NOT work for pdhgm() solver, so you'll have to choose an alternative explicitly when using that solver.

Exponential fits

For 1D data, we can use the expfit function to perform multi-exponential fits. We can use the function by specifying either the number of exponential components, or a vector which defines the starting points for the regression. See below:

NMRInversions.expfit — Function

expfit(input, n=1)

Fit an n-exponential function to the data in the input struct.

The outut is an ExpfitData structure.

Arguments:

input : 1D ExperimentData.

n : Integer specifying the number of exponential terms.

Optional arguments:

solver : Optim solver, default choice is IPNewton().

L : An integer specifying which norm of the residuals you want to minimize (default is 2).

The n argument can also be a vector of initial parameter guesses, in which case it also determines the number of exponential terms used. It should be of the form [a1, b1, a2, b2, ...], where a's are the amplitudes and b's are the parameters inside the exponentials.

The length of the vector in this case must be an even number.

The following examples might help to clarify:

expfit(input, [a,b]) -> mono-exponential fit with initial guess: a * exp.( (1/b) * x)

expfit(input, [a,b,c,d]) -> bi-exponential fit with initial guess: a * exp.( (1/b) * x) + c * exp.((1/d) * x)

expfit(input, [a,b,c,d,e,f]) -> tri-exponential fit with initial guess: a * exp.( (1/b) * x) + c * exp.((1/d) * x) + e * exp.((1/f) * x)

(where a,b,c,d,e,f are numbers of your choice)

Numbers of parameters beyond tri-exponential can also be used, but it is not recommended.

If you have some rough clue about what results you expect, it's best to define some starting points close to these. (especially important if you're using double or tri-exponential fits.)

Plotting

This package offers plotting capabilities, using its GLMakie extension. Simply use using GLMakie before or after using NMRInversions to load the package, and these functions become available.

We basically take the plot() function offered by GLMakie and extend it to types from the package.

You can use plot() for ExperimentData structures coming from import functions, e.g.:

data = import_csv(IR, "./path/to/file.csv")
plot(data)

Similarly, for inversion results:

data = import_csv(IR, "./path/to/file.csv")
results = invert(data)
plot(results)

Using a matrix or vector of InversionData structures (e.g. plot([results1, results2, results3]) will plot all of them on the same figure, for a quick comparison.

Data plots

Makie.plot — Method

plot(data::ExperimentData)

Quickly visualise the contents of a 1D NMR data structure.

Makie.plot — Method

plot(data::ExperimentData)

Interactive GUI to visualise the contents of a 2D NMR data structure.

1D inversion plots

Makie.plot — Method

plot(res::InversionData)

Run the GUI to plot the 1D inversion results and select peaks you want to label.

Makie.plot — Method

plot(res_mat::VecOrMat{InversionData};kwargs...)

Plot the results contained in a vector of InversionData structures.

The keyword (optional) arguments are:

selections : Whether to highlight the selections in each of the plots (default is true).

stacked : Whether to plot different curves as a stack, instead of superimposed. (default is true).

2D inversion plots

Makie.plot — Method

Makie.plot(res::InversionData{2}; kwargs...)

Run the GUI to plot the results and select peaks you want to label.

Keyword (optional) arguments:

dims : Dimensions of each plot (default: (400, 400)).

title : Title of the plot (default: "").

colormap : Color map of the plot (default: :viridis).

contf : Whether to use a filled contour plot (default : false).

levels : Number of contour levels (default: 40).

labelsizes : Size of the axis labels (default: (23, 23)).

ticksizes : Size of the axis ticks (default: (14, 14)).

titlesize : Size of the title (default: 17).

titlefont : Font of the title (default: :bold).

legendlabelsize : Size of the legend label (default: 12)

gap : The gap between the contour plot and the distribution plots (default: 0)

extend_grid : Whether to extend the gridlines to the plots on the sides (default: false)

legendposition : Where to put the legend (default: :lt )

bandplots : Whether to use band plots to highlight the selections in the top and right distributions (default: true)

tp: whether the map should be transposed

legend: whether the legend describing the selections should be displayed

Makie.plot — Method

plot(res_mat::AbstractVecOrMat{InversionData{2}}; kwargs...)

Plot the results contained in a matrix of InversionData{2} structures. The arrangement is determined by the shape of the matrix.

The arguments are:

res_mat : The InversionData{2} matrix or vector containing the fit results.

Keyword (optional) arguments:

dims : Dimensions of each plot (default: (400, 400)).

title : Title of the plot (default: "").

colormap : Color map of the plot (default: :viridis).

contf : Whether to use a filled contour plot (default : false).

levels : Number of contour levels (default: 40).

labelsizes : Size of the axis labels (default: (23, 23)).

ticksizes : Size of the axis ticks (default: (14, 14)).

titlesize : Size of the title (default: 17).

titlefont : Font of the title (default: :bold).

legendlabelsize : Size of the legend label (default: 12)

gap : The gap between the contour plot and the distribution plots (default: 0)

extend_grid : Whether to extend the gridlines to the plots on the sides (default: false)

legendposition : Where to put the legend (default: :lt )

bandplots : Whether to use band plots to highlight the selections in the top and right distributions (default: true)

tp: whether the map should be transposed

legend: whether the legend describing the selections should be displayed

Expfit plots

Makie.plot — Method

plot(res::expfit_struct...; kwargs...)

Plot the results of an expfit call.

This function can take any number of expfit_struct structures as input.

e.g. plot(data1) plots the data and the corresponding fit, but plot(data1, data2) or plot(data1, data2, data3) work as well, and they plot all of the results on the same plot.

Keyword (optional) arguments:

markersize : The size of the markers (default is 7).

normeq : Whether to plot the normalised form of the equation or not (default is false).

yscale : The scale of the y-axis (default is identity).

Miscellaneous

Weighted averages

Once you have selected some peaks in your inversion results through the GUI, you might want to extract the weighted averages of these selected peaks, to get the underlying relaxation times or diffusion coefficients. The following functions do the job:

NMRInversions.weighted_averages — Function

weighted_averages(r::InversionData{1})

Return a vector with the weighted averages for the selections in the input structure, and a vector with the respective area fractions of these selections.

weighted_averages(InversionData{2})

Return two vectors with the weighted averages for the selections in the input structure, one for each dimension, as well as a vector with the volume fractions of these selections.

For example:

results = invert(my_data)
plot(results)
"Select regions interactively through the GUI"
weighted_averages(results)