In order to take advantage of Julia's multiple dispatch, we need to define some structures which can be used as input to our functions.

Info

Structures are object which may contain multiple fields. The fields can be accessed using the dot syntax. For example, if we have a structure named foo, with two fields, a, and b, we can access the value of a using foo.a and the value of b using foo.b.

Pulse sequences - Data Axes

The following types are used so that the invert function can create the appropriate kernel for the inversion. Anywhere you see the seq keyword, one of the following types must be used.

NMRInversions.IR — Type

IR{T<:Real} <: DataAxis{T}

DataAxis type for inversion recovery measurements. The x field contains the corresponding time data.

NMRInversions.SR — Type

SR{T<:Real} <: DataAxis{T}

DataAxis type for saturation recovery measurements. The x field contains the corresponding time data.

NMRInversions.CPMG — Type

CPMG{T<:Real} <: DataAxis{T}

DataAxis type for CPMG measurements. The x field contains the corresponding time data.

NMRInversions.PFG — Type

PFG{T<:Real} <: DataAxis{T}

DataAxis type for pulsed field gradient measurements. The x field contains the corresponding b-factor data.

NMRInversions.Spectrum — Type

Spectrum{T<:Real} <: DataAxis{T}

DataAxis type for NMR spectra. The x field contains the corresponding ppm data.

NMRInversions.FID — Type

FID{T<:Real} <: DataAxis{T}

DataAxis type for free induction decay measurements. The x field contains the corresponding time data.

NMRInversions.FC — Type

FC{T<:Real} <: DataAxis{T}

DataAxis type for field cycling (or "dispersion") measurements. The x field contains the corresponding B0 data.

The types above represent just a special type of vector. Associating the x-axis of the data with a pulse sequence seems to be a fairly elegant way to make the package scale to n-dimensions, thus this convention is used from version 2.0 onwards.

Inversion solvers

These are used to let the invert function know which solver to use.

They can be used as input to the invert function as the 'solver' argument. (e.g., invert(data, solver=brd()) [3] or invert(data, solver=pdhgm(10,0.1)) [7]).

NMRInversions.brd — Type

brd()

Solver for tikhonov (L2) regularization, following this paper [4]. Very fast, but only uses the identity as tiknohonov matrix. It can be used as a "solver" for the invert function.

NMRInversions.pdhgm — Type

pdhgm(σ, τ, tol)

Primal dual hybrid gradient method for L1 regularization, following this paper [7].

It can be used as a "solver" for the invert function.

Positional (keyowrd) arguments:

sigma (default value 10)

tau (default value 0.1)

tol (default value 1e-5)

The particular choice of σ and τ is heuristic. The parameters σ and τ are step size parameters, which control convergence and stability of the algorithm. Convergence is guaranteed when τσ ≤ 1.

The best values of σ and τ will depend slightly on the scaling of the signal. Therefore, it is best to normalize the NMR signal to a maximum of 1 (this is default for the invert function, no need to do anything).

for convenience.

Note that for this method, the role of α is inverted, with larger α values leading to less smoothing, not more. For that reason, alpha=gcv() (Mitchell method) will not work. Please use alpha=gcv(starting_value) or `alpha=gcv(lowerlimit, upperlimit)', or some lcurve method instead.

NMRInversions.cdL1 — Type

cdL1(; iterations, tol)

Coordinate descent method for L1 regularization.

It can be used as a "solver" for the invert function.

Accepts two keyword (optional arguments):

iterations is an integer declaring the maximum amount of iterations before the algorithm stops (default value 10000).

tol is the relative tolerance (default value 1e-5)

NMRInversions.nnls — Type

nnls(; L, algorithm)

Non-negative least squares method for regularization, implemented using the NonNegLeastSquares package extension. Decent speed for small, 1D regularizations, but very slow for larger, 2D problems. It can be used as a "solver" for invert function.

L determines the tikhonov matrix.Can be either a matrix or an integer n. The latter will create a finite difference matrixof order n, which means that the penalty term will be the n'th derivative of the distribution.Defaults to 0, where L becomes the Identity matrix.

algorithm is one of:

:nnls

:fnnls

:pivot

NMRInversions.optim_nnls — Type

optim_nnls(; L, algorithm, opts)

Simple non-negative least squares method for regularization, implemented using Optim.jl. All around effective, but can be slow for large problems, such as 2D inversions. It can be used as a "solver" for invert function.

L determines the tikhonov matrix.Can be either a matrix or an integer n. The latter will create a finite difference matrixof order n, which means that the penalty term will be the n'th derivative of the distribution.Defaults to 0, where L becomes the Identity matrix.

algorithm is of Optim.FirstOrderOptimizer type (defaults to LBFGS()).

opts an Optim.Options() structure which can provide some preferences to the solver.

The following are also implemented through external package extensions.

NMRInversions.jump_nnls — Type

jump_nnls(L, solver)

Jump non-negative least squares method for tikhonov (L2) regularization, implemented using the JuMP package extension. All around effective, but can be slow for large problems, such as 2D inversions, unless a powerful solver like gurobi is used.

L determines the tikhonov matrix.Can be either a matrix or an integer n. The latter will create a finite difference matrixof order n, which means that the penalty term will be the n'th derivative of the distribution.Defaults to 0, where L becomes the Identity matrix.

solver is passed directly into JuMP, see its documentation for available options.

This one is still experimental and not well-tested, please submit an issue if you encounter any difficulties.

Finding optimal alpha

These are methods for finding the optimal regularization parameter. They can be used as input to the invert function as the 'alpha' argument (e.g., invert(data, alpha=gcv()) or invert(data, alpha=lcurve(0.001,1,64)) ). If you'd like to use a particular value of alpha, you can just use that number instead (invert(data, alpha=1)).

NMRInversions.gcv — Function

gcv()

Constructor for the gcv method described in Mitchell 2012. No arguments required.

gcv(lower, upper ; kwargs...)

Constructor for finding the optimal alpha value via gcv score univariate optimization, given some lower and upper bounds.

Necessary (positional) arguments:

lower is the lower bound, or lowest alpha value the algorighm will consider.

upper is the upper bound, or highest alpha value the algorighm will consider.

Optional (keyword) arguments:

algorithm determines which method will be used by Optim.jl to solve the problem.

Default is Brent() and the only alternative is GoldenSection(), which is normally slower.

abs_tol determines what's the smallest absolute change in the gcv score the algorithm should care about.

Default is 1e-3.

rel_tol determines what's the smallest relative change in the gcv score the algorithm should care about.

Default is 1e-1.

gcv(start; kwargs...)

Constructor for finding the optimal alpha value via gcv score box optimization, given a starting value.

Necessary (positional) arguments:

start is the starting alpha value.

Choose something sensible, usually a value between 0.1 and 10 would work well.

Optional (keyword) arguments:

algorithm determines which method will be used by Optim.jl to solve the problem.

Default is LBFGS(). Only first order optimizers can be chosen here. For more details, refer to Optim.jl documentation.

opts an Optim.Options() structure which can provide some preferences to the solver.

NMRInversions.lcurve — Function

lcurve(lower, upper ; kwargs...)

Constructor for finding the optimal alpha value via lcurve curvature univariate optimization, given some lower and upper bounds.

Necessary (positional) arguments:

lower is the lower bound, or lowest alpha value the algorighm will consider.

upper is the upper bound, or highest alpha value the algorighm will consider.

Optional (keyword) arguments:

algorithm determines which method will be used by Optim.jl to solve the problem.

Default is Brent() and the only alternative is GoldenSection(), which is normally slower.

abs_tol determines what's the smallest absolute change in the lcurve score the algorithm should care about.

Default is 1e-3.

rel_tol determines what's the smallest relative change in the lcurve score the algorithm should care about.

Default is 1e-1.

lcurve(start; kwargs...)

Constructor for finding the optimal alpha value via lcurve curvature box optimization, given a starting value.

Necessary (positional) arguments:

start is the starting alpha value.

Choose something sensible, usually a value between 0.1 and 10 would work well.

Optional (keyword) arguments:

algorithm determines which method will be used by Optim.jl to solve the problem.

Default is LBFGS(). Only first order optimizers can be chosen here. For more details, refer to Optim.jl documentation.

opts an Optim.Options() structure which can provide some preferences to the solver.

Please have a look here.

lcurve(lowest_value, highest_value, number_of_steps)

Constructor for testing all L-curve curvatures between some bounds and picking the optimal.

lowest_value is the lowest alpha value.

highest_value is the highest alpha value.

number_of_steps is the number of alpha values that will be tested.

This is a very crude and rather slow method, mostly for demonstration purposes.

Inversion and expfit inputs

These can be used as inputs containing all of the necessary information to run the invert and expfit functions (e.g. if a is an ExperimentData object, invert(a) does the job).

NMRInversions.ExperimentData — Type

ExperimentData(axes, data)

Struct containing NMR data.

axes should be a tuple of DataAxiss with the x data in each of them.

data should be an aray of n-dimensions with the recorded data.

SNR is the signal-to-noise ratio (calculated automatically for complex data)

W is a tuple containing the precision matrix (inverse of the covariance

matrix) in each of the dimensions. Relevant for compressed data, identity by default.

Inversion and expfit outputs

When you run an inversion function (e.g. r = invert(a)), the output is an InversionData structure.

NMRInversions.InversionData — Type

InversionData()

Output of the invert function, containing all the relevant information:

input : ExperimentData structure that was used as input to the inversion function.

axes : "x-axes" of the output (i.e., relaxation times, diffusion coefficients).

data : the results of the inversion.

residuals : an array containing the residuals of the fit.

alpha : the smooting term used to produce the data array.

filter : an array working as a mask to the data array (mostly for visualisation).

selections : polygon selections for the data array (mostly for visualisation).

title : a title descibing this object (mostly for visualisation).

NMRInversions.ExpfitData — Type

Output of the expfit function. Structure containing information about multiexponential fits.

The fields are as follows:

u : The fitted parameters for the mexp function.

u0 : The initial parameters for the mexp function.

residuals : The residuals of the fit.

eq : The equation of the fitted function, printed as a string.

eqn : Same as above, normalised.

title : A title describing the data.

Kernel structure

NMRInversions.svd_kernel_struct — Type

svd_kernel_struct(K,g,U,S,V)

A structure containing the following fields:

K, the kernel matrix.

g, the data vector.

U, the left singular values matrix.

S, the singular values vector.

V, the right singular values matrix.

To access the fields of a structure, we use the dot notation, e.g. if the structure is called a and we want to access the kernel contained in there, we type a.K