Intro

This page contains the documentation for various useful functions in the NMRInversions package.

Tip

From the julia command line, you can enter '?', followed by the name of any function, struct, or object you want to learn more about (try it!). After typing using NMRInversions in the julia console, this feature will work for all the functions mentioned below.

Info

In julia, function definitions look like this:

foo(x, y, z ; a, b)

For the example above, foo is the name of the function, and the contents of the parenteses are the arguments. Within the parenteses, we got two types of arguments:

Positional arguments

Everything that appears before the semicolon ; (x,y and z in this example) is necessary, and must be given in a specific order.

Keyword arguments.

Everything that appears after the semicolon ; (b and a in this example) is optional, and can be given in any order, but its name must be specified.

In the example above, we can call the function by typing foo(1, 2, 3), (in which case, x=1, y=2, z=3, and the default value for a and b will be used). You can also call the function by typing foo(1, 2, 3, a=3), to specify the value for a, or by typing foo(1, 2, 3, b=3, a = 2), to specify the value for both a and b.

Sometimes, if there are many keyword arguments, we write the function as foo(x, y ; kwargs...). For the necessary arguments before the semicolon, order matters. For the keyword arguments after the semicolon, order does not matter, but the name of each argument must be specified. For more information, please refer to this link.

Importing data functions

This package offers some functions to import NMR experiment data of various formats. Alternatively, you can of course import your data however you see fit. If a format you're working with is not yet supported, please submit an issue and we'll work on it.

The most basic use case would be using data saved in a csv format, where there are only two columns, one for your x-axis (time for relaxation and b-factor for diffusion) and one for your y-axis (signal intensity).

NMRInversions.import_csv — Method

import_csv(seq, file)

Import data from a CSV file. The function reads the file and returns an input1D structure.

seq is the 1D pulse sequence (e.g. IR, CPMG, PFG)

file is the path to the CSV file which contains the data (x, y) in two respective columns.

The function can be called without the seq argument, and the output will be the x and y vectors (use it like, x,y =import_csv()). Alternatively, the function can also be called with only the seq argument, in which case a file dialog will open to select the file (use it like, data = import_csv(IR))

Please note that this function will just import the data as is, without any unit conversions. Ensure that your x-axis is in SI.

If you're using a spinsolve instrument, you can use the import_spinsolve function. This one requires two files as an input. The aqcu.par is automatically exported by SpinsolveExpert, but you might have to export your data file manually in a csv format.

NMRInversions.import_spinsolve — Method

import_spinsolve(files)

Import data from a Spinsolve experiment. Two paths must be provided as follows (order is not important):

files = [.../datafile.csv , .../acqu.par]

Calling this function without an argument by typing import_spinsolve() will open a file dialog to select the files. The function reads the acqu.par.bak file to get the acquisition parameters, and the .dat file to get the data. The function returns an input2D structure.

For geospec instruments, you can export your raw data as a text file. That text file can be read by the import_geospec function.

NMRInversions.import_geospec — Method

import_geospec(dir)

Import data from a .txt format, as exported by Geospec instruments.

The function reads the relevant information, performs a phase correction on the data, and returns an input1D or input2D structure.

Calling this function without an argument by typing import_geospec() will open a file dialog to select the .txt file.

Inversion functions

The most important function is invert(), which is the main function of the package. It works as follows:

NMRInversions.invert — Method

Inversion for 1D pulse sequences:

invert(seq, x, y ; lims, alpha, solver)

This function will build a kernel and use it to perform an inversion using the algorithm of your choice. The output is an inv_out_1D structure.

Necessary (positional) arguments:

seq is the 1D pulse sequence (e.g. IR, CPMG, PFG)

x is the experiment x axis (time or b factor etc.)

y is the experiment y axis (intensity of the NMR signal)

Optional (keyword) arguments:

lims=(a,b,c) will set the "limits" of the output X,

so that it starts from 10^a, ends in 10^b and consists of c logarithmically spaced values (default is (-5, 1, 128) for relaxation and (-10, -7, 128) for diffusion). Alternatiively, a vector of values can be used directly, if more freedom is needed (e.g. lims=exp10.(range(-4, 1, 128))).

alpha determines the smoothing term. Use a real number for a fixed alpha. No selection will lead to automatically determining alpha through the defeault method, which is gcv.

solver is the algorithm used to do the inversion math. Default is brd.

normalize will normalize y to 1 at the max value of y. Default is true.

NMRInversions.invert — Method

invert(data::input1D ; kwargs...)

Instead of the positional arguments seq, x and y, you can use a single input1D structure, which contains the same information. Especially useful if you're using the output of one of the import functions (look documentation tutorial section).

Due to julia's multiple dispatch, it is possible to define a function with the same name but different arguments, to achieve different results.

Because of that, the inversion function also works for 2D inversions, if the following arguments are used instead:

NMRInversions.invert — Method

Inversion for 2D pulse sequences:

invert(seq, x_direct, x_indirect, X_direct, X_indirect, Data; lims1, lims2, alpha, solver)

This function will build a kernel and use it to perform an inversion using the algorithm of your choice. The output is an inv_out_2D structure.

Necessary (positional) arguments:

seq is the 2D pulse sequence (e.g. IRCPMG)

x_direct is the direct dimension acquisition parameter (e.g. the times when you aquire CPMG echoes).

x_indirect is the indirect dimension acquisition parameter (e.g. all the delay times τ in your IR sequence).

Data is the 2D data matrix of complex data.

Optional (keyword) arguments:

lims1 determines the output "range" of the inversion in the direct dimension (e.g. T₂ times in IRCPMG)

lims2 determines the output "range" of the inversion in the indirect dimension (e.g. T₁ times in IRCPMG)

In both cases above, you can use a tuple specifying the limits of the range, or a vector of values, same as the lims argument in the 1D inversion.

alpha determines the smoothing term. Use a real number for a fixed alpha. No selection will lead to automatically determining alpha through the defeault method, which is gcv.

solver is the algorithm used to do the inversion math. Default is brd.

normalize will normalize Data to 1 at the max value of Data. Default is true.

NMRInversions.invert — Method

invert(data::input2D ; kwargs...)

Instead of the positional arguments seq, x_direct , x_indirect and Data, you can use a single input2D structure, which contains the same information. Especially useful if you're using the output of one of the import functions (look documentation tutorial section).

Exponential fit functions

For 1D data, we can use the expfit function to perform multiexponential fits. We can use the function by specifying either the number of exponential components, or a vector which defines the starting points for the regression. See below:

NMRInversions.expfit — Method

expfit(n, seq, x, y; kwargs...)

Fit an n-exponential function to the data x and y.

The outut is an expfit_struct structure.

Arguments:

n : Integer specifying the number of exponential terms.

seq : pulse sequence.

x : acquisition x parameter (time for relaxation or b-factor for diffusion).

y : acquisition y parameter (magnetization).

Optional arguments:

solver : OptimizationOptimJL solver, defeault choice is BFGS().

normalize : Normalize the data before fitting? (default is true).

The n argument can also be a vector of initial parameter guesses, in which case it also determines the number of exponential terms used. It should be of the form [a1, b1, a2, b2, ...], where a's are the amplitudes and b's are the parameters inside the exponentials.

The length of the vector in this case must be an even number.

The following examples might help to clarify:

expfit([a,b] , CPMG, x, y) -> mono-exponential fit with initial guess: a * exp.( (1/b) * x)

expfit([a,b,c,d] , CPMG, x, y) -> bi-exponential fit with initial guess: a * exp.( (1/b) * x) + c * exp.((1/d) * x)

expfit([a,b,c,d,e,f] , CPMG, x, y) -> tri-exponential fit with initial guess: a * exp.( (1/b) * x) + c * exp.((1/d) * x) + e * exp.((1/f) * x)

(where a,b,c,d,e,f are numbers of your choice)

Numbers of parameters beyond tri-exponential can also be used, but it is not recommended.

NMRInversions.expfit — Method

expfit(n, data; kwargs...)

Similar to the invert fucntion, expfit can be called using an input1D structure.

Arguments:

n : Integer specifying the number of exponential terms.

data : input1D structure containing the data to be fitted.

If you have some rough clue about what results you expect, it's best to define some starting points close to these. (especially important if you're using double or tri-exponential fits.)

Plotting functions

This package offers plotting capabilities, using its GLMakie extension. Simply use using GLMakie before or after using NMRInversions to load the package, and these functions become available.

We basically take the plot() function offered by GLMakie and extend it to types from the package.

For example, for 1D inversions, we have:

MakieCore.plot — Method

plot(res::NMRInversions.inv_out_1D...; kwargs...)

Plot the results contained in a inv_out_1D structure. This function can take any number of inv_out_1D structures as input.

MakieCore.plot! — Method

plot!(fig, res...; title)

Plot the results contained in a inv_out_1D structure on a figure or a grid position object. This function can take any number of inv_out_1D structures as input.

The arguments are:

fig : The figure or grid position object.

res : One or more inv_out_1D structures containing the fit results.

Keyword (optional) arguments:

title : Title of the plot.

For expfits, we have:

MakieCore.plot — Method

plot(res::expfit_struct...; kwargs...)

Plot the results of an expfit call.

This function can take any number of expfit_struct structures as input.

e.g. plot(data1) plots the data and the corresponding fit, but plot(data1, data2) or plot(data1, data2, data3) work as well, and they plot all of the results on the same plot.

This function creates a figure, and calls plot!(fig, res) for each of the results in res, so please refer to the plot! documentation for more information.

MakieCore.plot! — Method

plot!(fig, res...; names, markersize, normeq)

Plots the results of an expfit call on a figure or a grid position object.

The arguments are:

fig : The figure or grid position object.

res : One or more expfit_struct structures containing the fit results.

Keyword (optional) arguments:

markersize : The size of the markers (default is 7).

normeq : Whether to plot the normalised form of the equation or not (default is true).

Note that the res inputs are not a vector, but individual expfit_struct structures, like: plot!(fig, data1, data2, data3). If you want to use a vector of expfit_struct structures, make sure to splat it by using ... in the function call (e.g. plot!(fig, [data1, data2, data3]...)).

And for 2D inversions, there's an interactive gui to characterize the inversion resutls:

MakieCore.plot — Method

Makie.plot(res::inv_out_2D)

Run the GUI to plot the results and select peaks you want to label.

MakieCore.plot — Method

plot(results::AbstractVecOrMat{inv_out_2D}; kwargs...)

Plot the results contained in a matrix of inv_out_2D structures. The arrangement is determined by the shape of the matrix.

The arguments are:

results : The inv_out_2D matrix or vector containing the fit results.

Keyword (optional) arguments:

title : Title of the plot (default: "").

colormap : Color map of the plot (default: :viridis).

contf : Whether to use a filled contour plot (default: false).

levels : Number of contour levels (default: 40).

Basically, if your inv_out_2D structure is a variable named results, you can run the GUI by running plots(results), or you can create a simple plot by running plot([results]). The latter can be used with matrices of inv_out_2D structures, and the layout of the resulting plot will reflect the layout of the matrix.