This page contains the documentation for various useful functions in the NMRInversions package.

Tip

From the Julia command line, you can enter '?', followed by the name of any function, struct, or object you want to learn more about (try it!). After typing using NMRInversions in the Julia console, this feature will work for all the functions mentioned below.

Info

In Julia, function definitions look like this:

foo(x, y, z ; a, b)

For the example above, foo is the name of the function, and the contents of the parenteses are the arguments. Within the parenteses, we got two types of arguments:

Positional arguments

Everything that appears before the semicolon ; (x,y and z in this example) is necessary, and must be given in a specific order.

Keyword arguments.

Everything that appears after the semicolon ; (b and a in this example) is optional, and can be given in any order, but its name must be specified.

In the example above, we can call the function by typing foo(1, 2, 3), (in which case, x=1, y=2, z=3, and the default value for a and b will be used). You can also call the function by typing foo(1, 2, 3, a=3), to specify the value for a, or by typing foo(1, 2, 3, b=3, a = 2), to specify the value for both a and b.

Sometimes, if there are many keyword arguments, we write the function as foo(x, y ; kwargs...). For the necessary arguments before the semicolon, order matters. For the keyword arguments after the semicolon, order does not matter, but the name of each argument must be specified. For more information, please refer to this link.

Importing data

This package offers some functions to import NMR experiment data of various formats. Alternatively, you can of course import your data however you see fit. If a format you're working with is not yet supported, please submit an issue and we'll work on it.

The most basic use case would be using data saved in a csv format, where there are only two columns, one for your x-axis (time for relaxation and b-factor for diffusion) and one for your y-axis (signal intensity).

NMRInversions.import_csv — Method

import_csv(seq, file)

Import data from a CSV file. The function reads the file and returns an input1D structure.

seq is the 1D pulse sequence (e.g. IR, CPMG, PFG)

file is the path to the CSV file which contains the data (x, y) in two respective columns.

The function can be called without the seq argument, and the output will be the x and y vectors (use it like, x,y =import_csv()). Alternatively, the function can also be called with only the seq argument, in which case a file dialog will open to select the file (use it like, data = import_csv(IR))

Please note that this function will just import the data as is, without any unit conversions. Ensure that your x-axis is in seconds for relaxation, or in seconds/metres² * 1e-9 for PFG experiments.

If you're using a spinsolve instrument, you can use the import_spinsolve function. This one requires two files as an input. The aqcu.par is automatically exported by SpinsolveExpert, but you might have to export your data file manually in a csv format.

NMRInversions.import_spinsolve — Method

import_spinsolve(files)

Import data from a Spinsolve expert experiment. Two paths must be provided as follows (order is not important):

files = [.../datafile.csv , .../acqu.par]

Calling this function without an argument by typing import_spinsolve() will open a file dialogue to select the files.

The function reads the acqu.par.bak file to get the acquisition parameters, and the .dat file to get the data. The function returns an input2D structure.

For geospec instruments, you can export your raw data as a text file. That text file can be read by the import_geospec function.

NMRInversions.import_geospec — Method

import_geospec(dir)

Import data from a .txt format, as exported by Geospec instruments.

The function reads the relevant information, performs a phase correction on the data, and returns an input1D or input2D structure.

Calling this function without an argument by typing import_geospec() will open a file dialog to select the .txt file.

For.tnt files from Tecmag consoles, import_tnt can be used. It should be noted however, that since pulse sequence implementations might vary, automatic parsing of the data into a input1D or input2D struct might fail. Therefore, the more manual read_tnt_data and read_tnt_header functions can be used for fine-tuned cases. If that sounds confusing, please submit an issue and we'll work on it!

NMRInversions.import_tnt — Method

import_tnt(seq, filename)

Read data from a tecmag .tnt file, and return a input1D or input2D object with all the relevant information.

This function is experimental, and since tecmag consoles work mainly with custom pulse sequences, it might not cover some cases. If it does not work for you, please submit an issue.

Necessary (positional) arguments:

seq is the pulse sequence (e.g. IR, CPMG, PFG)

filename is the path of the file containing the data.

Optional (keyword) arguments:

echotime is the string that is used to denote the Echo time as

defined in the pulse sequence. Default value is "Echo_Time".

Calling this function without a filename argument, e.g.: import_tnt(seq), will open a file dialogue to select the .tnt file.

NMRInversions.read_tnt_data — Method

read_tnt_data(filename::String)

Read raw data from a tecmag .tnt file and return a complex vector containing all the values.

NMRInversions.read_tnt_header — Method

read_tnt_header(filename::String)

Read the header of a .tnt file and return a dictionary with all the relevant information.

Inversion

The most important function is invert(), which is the main function of the package. It works as follows:

NMRInversions.invert — Method

Inversion for 1D pulse sequences:

invert(seq, x, y ; lims, alpha, solver, normalize)

This function will build a kernel and use it to perform an inversion using the algorithm of your choice. The output is an inv_out_1D structure.

Necessary (positional) arguments:

seq is the 1D pulse sequence (e.g. IR, CPMG, PFG)

x is the experiment x axis (time or b factor etc.)

y is the experiment y axis (intensity of the NMR signal)

Optional (keyword) arguments:

lims=(a,b,c) will set the "limits" of the output X,

so that it starts from 10^a, ends in 10^b and consists of c logarithmically spaced values. Alternatively, a vector of values can be used directly, if more freedom is needed (e.g. lims=exp10.(range(-4, 1, 128))). By default, 128 points are used, logarithmically spaced between the first value in x divided by 7, and the final value in x multiplied by 7.

alpha determines the smoothing term. Use a real number for a fixed alpha,

or look at the "Types and Structures - Finding Optimal Alpha" section of the documentation for other options. No selection will lead to automatically determining alpha through the default method, which is gcv().

solver is the algorithm used to do the inversion math. Default is brd().

normalize will normalize y to 1 at the max value of y. Default is true.

silent will suppress the output of the alpha search. Default is false.

NMRInversions.invert — Method

invert(data::input1D ; kwargs...)

Instead of the positional arguments seq, x and y, you can use a single input1D structure, which contains the same information. Especially useful if you're using the output of one of the import functions (look documentation tutorial section).

Due to Julia's multiple dispatch, it is possible to define a function with the same name but different arguments, to achieve different results.

Because of that, the inversion function also works for 2D inversions, if the following arguments are used instead:

NMRInversions.invert — Method

Inversion for 2D pulse sequences:

invert(seq, x_direct, x_indirect, Data; lims1, lims2, alpha, solver, normalize)

This function will build a kernel and use it to perform an inversion using the algorithm of your choice. The output is an inv_out_2D structure.

Necessary (positional) arguments:

seq is the 2D pulse sequence (e.g. IRCPMG)

x_direct is the direct dimension acquisition parameter (e.g. the times when you aquire CPMG echoes).

x_indirect is the indirect dimension acquisition parameter (e.g. all the delay times τ in your IR sequence).

Data is the 2D data matrix of complex data.

Optional (keyword) arguments:

lims1 determines the output "range" of the inversion in the direct dimension (e.g. T₂ times in IRCPMG)

lims2 determines the output "range" of the inversion in the indirect dimension (e.g. T₁ times in IRCPMG)

In both cases above, you can use a tuple specifying the limits of the range, or a vector of values, same as the lims argument in the 1D inversion function. By default, 64 points are used, logarithmically spaced between the first value in x divided by 7, and the final value in x multiplied by 7 (for both direct and indirect x).

alpha determines the smoothing term. Use a real number for a fixed alpha. No selection will lead to automatically determining alpha through the default method, which is gcv.

solver is the algorithm used to do the inversion math. Default is brd().

normalize will normalize Data so that its maximum value is 1. Default is true.

silent will suppress the output of the alpha search. Default is false.

NMRInversions.invert — Method

invert(data::input2D ; kwargs...)

Instead of the positional arguments seq, x_direct , x_indirect and Data, you can use a single input2D structure, which contains the same information. Especially useful if you're using the output of one of the import functions (look documentation tutorial section).

Finding alpha

Here we provide two options for finding the optimal value for alpha, namely Generalized Cross Validation (GCV) or L-curve. Generally gcv seems slightly more reliable in NMR, but it's far from perfect, so it's good to have alternatives and cross-check. The following methods can be used as inputs for the alpha argument in the invert function:

NMRInversions.gcv — Method

gcv()

Constructor for the gcv method described in Mitchell 2012. No arguments required.

NMRInversions.gcv — Method

gcv(start; kwargs...)

Constructor for finding the optimal alpha value via gcv score box optimization, given a starting value.

Necessary (positional) arguments:

start is the starting alpha value.

Choose something sensible, usually a value between 0.1 and 10 would work well.

Optional (keyword) arguments:

algorithm determines which method will be used by Optim.jl to solve the problem.

Default is LBFGS(). Only first order optimizers can be chosen here. For more details, refer to Optim.jl documentation.

opts an Optim.Options() structure which can provide some preferences to the solver.

Please have a look here.

NMRInversions.gcv — Method

gcv(lower, upper ; kwargs...)

Constructor for finding the optimal alpha value via gcv score univariate optimization, given some lower and upper bounds.

Necessary (positional) arguments:

lower is the lower bound, or lowest alpha value the algorighm will consider.

upper is the upper bound, or highest alpha value the algorighm will consider.

Optional (keyword) arguments:

algorithm determines which method will be used by Optim.jl to solve the problem.

Default is Brent() and the only alternative is GoldenSection(), which is normally slower.

abs_tol determines what's the smallest change in alpha the algorithm should care about.

Default is 1e-3.

NMRInversions.lcurve — Method

lcurve(lowest_value, highest_value, number_of_steps)

Constructor for testing all L-curve curvatures between some bounds and picking the optimal.

lowest_value is the lowest alpha value.

highest_value is the highest alpha value.

number_of_steps is the number of alpha values that will be tested.

This is a very crude and rather slow method, mostly for demonstration purposes.

NMRInversions.lcurve — Method

lcurve(start; kwargs...)

Constructor for finding the optimal alpha value via lcurve curvature box optimization, given a starting value.

Necessary (positional) arguments:

start is the starting alpha value.

Choose something sensible, usually a value between 0.1 and 10 would work well.

Optional (keyword) arguments:

algorithm determines which method will be used by Optim.jl to solve the problem.

Default is LBFGS(). Only first order optimizers can be chosen here. For more details, refer to Optim.jl documentation.

opts an Optim.Options() structure which can provide some preferences to the solver.

Please have a look here.

NMRInversions.lcurve — Method

lcurve(lower, upper ; kwargs...)

Constructor for finding the optimal alpha value via lcurve curvature univariate optimization, given some lower and upper bounds.

Necessary (positional) arguments:

lower is the lower bound, or lowest alpha value the algorighm will consider.

upper is the upper bound, or highest alpha value the algorighm will consider.

Optional (keyword) arguments:

algorithm determines which method will be used by Optim.jl to solve the problem.

Default is Brent() and the only alternative is GoldenSection(), which is normally slower.

abs_tol determines what's the smallest change in alpha the algorithm should care about.

Default is 1e-3.

The gcv() method [2] usually involves the least amount of function calls and it is thus much faster, thus used as the default option. If you want more precision, the univariate or box methods should be used instead. Note that gcv() will NOT work for pdhgm() solver, so you'll have to choose an alternative explicitly when using that solver.

Exponential fits

For 1D data, we can use the expfit function to perform multiexponential fits. We can use the function by specifying either the number of exponential components, or a vector which defines the starting points for the regression. See below:

NMRInversions.expfit — Method

expfit(n, seq, x, y; kwargs...)

Fit an n-exponential function to the data x and y.

The outut is an expfit_struct structure.

Arguments:

n : Integer specifying the number of exponential terms.

seq : pulse sequence.

x : acquisition x parameter (time for relaxation or b-factor for diffusion).

y : acquisition y parameter (magnetization).

Optional arguments:

solver : Optim solver, default choice is IPNewton().

normalize : Normalize the data before fitting? (default is true).

L : An integer specifying which norm of the residuals you want to minimize (default is 2).

The n argument can also be a vector of initial parameter guesses, in which case it also determines the number of exponential terms used. It should be of the form [a1, b1, a2, b2, ...], where a's are the amplitudes and b's are the parameters inside the exponentials.

The length of the vector in this case must be an even number.

The following examples might help to clarify:

expfit([a,b] , CPMG, x, y) -> mono-exponential fit with initial guess: a * exp.( (1/b) * x)

expfit([a,b,c,d] , CPMG, x, y) -> bi-exponential fit with initial guess: a * exp.( (1/b) * x) + c * exp.((1/d) * x)

expfit([a,b,c,d,e,f] , CPMG, x, y) -> tri-exponential fit with initial guess: a * exp.( (1/b) * x) + c * exp.((1/d) * x) + e * exp.((1/f) * x)

(where a,b,c,d,e,f are numbers of your choice)

Numbers of parameters beyond tri-exponential can also be used, but it is not recommended.

NMRInversions.expfit — Method

expfit(n, data; kwargs...)

Similar to the invert function, expfit can be called using an input1D structure.

Arguments:

n : Integer specifying the number of exponential terms.

data : input1D structure containing the data to be fitted.

If you have some rough clue about what results you expect, it's best to define some starting points close to these. (especially important if you're using double or tri-exponential fits.)

Plotting

This package offers plotting capabilities, using its GLMakie extension. Simply use using GLMakie before or after using NMRInversions to load the package, and these functions become available.

We basically take the plot() function offered by GLMakie and extend it to types from the package.

You can use plot() for input1D or input2D structures coming from import functions, e.g.:

data = import_csv(IR, "./path/to/file.csv")
plot(data)

Similarly, for inversion results:

data = import_csv(IR, "./path/to/file.csv")
results = invert(data)
plot(results)

Using a matrix or vector of inv_out_1D or inv_out_2D structures (e.g. plot([results1, results2, results3]) will plot all of them on the same figure, for a quick comparison.

Data plots

Makie.plot — Method

plot(data::input1D)

Quickly visualise the contents of a 1D NMR data structure.

Makie.plot — Method

plot(data::input2D)

Interactive GUI to visualise the contents of a 2D NMR data structure.

1D inversion plots

Makie.plot — Method

plot(res::inv_out_1D)

Run the GUI to plot the 1D inversion results and select peaks you want to label.

Makie.plot — Method

plot(res_mat::VecOrMat{inv_out_1D};kwargs...)

Plot the results contained in a vector of inv_out_1D structures.

The keyword (optional) arguments are:

selections : Whether to highlight the selections in the plots (default is false).

2D inversion plots

Makie.plot — Method

Makie.plot(res::inv_out_2D)

Run the GUI to plot the results and select peaks you want to label.

Makie.plot — Method

plot(results::AbstractVecOrMat{inv_out_2D}; kwargs...)

Plot the results contained in a matrix of inv_out_2D structures. The arrangement is determined by the shape of the matrix.

The arguments are:

results : The inv_out_2D matrix or vector containing the fit results.

Keyword (optional) arguments:

dims : Dimensions of each plot (default: (400, 400)).

title : Title of the plot (default: "").

colormap : Color map of the plot (default: :viridis).

contf : Whether to use a filled contour plot (default : false).

levels : Number of contour levels (default: 40).

labelsizes : Size of the axis labels (default: (23, 23)).

ticksizes : Size of the axis ticks (default: (14, 14)).

titlesize : Size of the title (default: 17).

titlefont : Font of the title (default: :bold).

legendlabelsize : Size of the legend label (default: 12)

gap : The gap between the contour plot and the distribution plots (default: 0)

extend_grid : Whether to extend the gridlines to the plots on the sides (default: false)

legendposition : Where to put the legend (default: :lt )

bandplots : Whether to use band plots to highlight the selections in the top and right distributions (default: true)

Expfit plots

Makie.plot — Method

plot(res::expfit_struct...; kwargs...)

Plot the results of an expfit call.

This function can take any number of expfit_struct structures as input.

e.g. plot(data1) plots the data and the corresponding fit, but plot(data1, data2) or plot(data1, data2, data3) work as well, and they plot all of the results on the same plot.

Keyword (optional) arguments:

markersize : The size of the markers (default is 7).

normeq : Whether to plot the normalised form of the equation or not (default is true).

yscale : The scale of the y-axis (default is identity).

Makie.plot! — Method

plot!(fig, res...; names, markersize, normeq)

Plots the results of an expfit call on a figure or a grid position object.

The arguments are:

fig : The figure or grid position object.

res : One or more expfit_struct structures containing the fit results.

Keyword (optional) arguments:

markersize : The size of the markers (default is 7).

normeq : Whether to plot the normalised form of the equation or not (default is true).

yscale : The scale of the y-axis (default is identity).

Note that the res inputs are not a vector, but individual expfit_struct structures, like: plot!(fig, data1, data2, data3). If you want to use a vector of expfit_struct structures, make sure to splat it by using ... in the function call (e.g. plot!(fig, [data1, data2, data3]...)).

Miscellaneous

Weighted averages

Once you have selected some peaks in your inversion results through the GUI, you might want to extract the weighted averages of these selected peaks, to get the underlying relaxation times or diffusion coefficients. The following functions do the job:

NMRInversions.weighted_averages — Method

weighted_averages(r::inv_out_1D)

Return a vector with the weighted averages for the selections in the input structure, and a vector with the respective area fractions of these selections.

NMRInversions.weighted_averages — Method

weighted_averages(r::inv_out_2D)

Return two vectors with the weighted averages for the selections in the input structure, one for each dimension, as well as a vector with the volume fractions of these selections.

For example:

results = invert(my_data)
plot(results)
"Select regions interactively through the GUI"
weighted_averages(results)

Trimming data

If you want a "quick and dirty" way to exclude some data points from your imported data, you may use the 'trim' function.

NMRInversions.trim — Method

trim(::input1D, ::Int=0, ::Int=0)

Return an input1D structure which excludes a selected amount of points from the beginning and the end of the x and y arrays.

For example, trim(data,2,3) will remove the first 2 and the last 3 data points.

NMRInversions.trim — Method

trim(::input2D; direct::Tuple{Int,Int}=(0,0), indirect::Tuple{Int,Int}=(0,0))

Return an input2D structure which excludes a selected amount of points from the beginning and the end of the direct and indirect dimensions of the data matrix, respectively.

For example, trim(data, direct=(2,3), indirect=(1,4)) will remove the first 2 and the 3 last columns of the data matrix, as well as the first row and the last four rows.

That means, for T1T2 data, you will remove the first 2 and last 3 echo points for T2 (direct dimension), and the first 1 and last 4 inversion recovery points (indirect dimension).

You can use the output of trim() directly as the inversion input:

invert(trim(import_csv(CPMG,"/path/to/data") , 3 ))

The one-liner above will remove the first three data points from the data and pass that into the inversion function.