Types and Structures

In order to take advantage of julia's multiple dispatch, we need to define some structures which can be used as input to our functions.

Info

Structures are object which may contain multiple fields. The fields can be accessed using the dot syntax. For example, if we have a structure named foo, with two fields, a, and b, we can access the value of a using foo.a and the value of b using foo.b.

Pulse sequences

The following types are used so that the invert function can create the appropriate kernel for the inversion. Anywhere you see the seq keyword, one of the following must be used.

NMRInversions.IR — Type

Inversion recovery pulse sequence for 1D relaxation experiments. It can be used wherever the seq argument is required.

NMRInversions.SR — Type

Saturation recovery pulse sequence for 1D relaxation experiments. It can be used wherever the seq argument is required.

NMRInversions.CPMG — Type

CPMG pulse sequence for 1D relaxation experiments. It can be used wherever the seq argument is required.

NMRInversions.PFG — Type

Pulsed field gradient pulse sequence for 1D diffusion experiments. It can be used wherever the seq argument is required.

NMRInversions.IRCPMG — Type

Inversion recovery - CPMG pulse sequence for 2D relaxation experiments (T1-T2). The direct dimension is the T2, and the indirect dimension is the T1 acquisition times. It can be used wherever the seq argument is required.

Inversion solvers

These are used to let the invert fucntion know which solver to use. They can be used as input to the invert function as the 'solver' argument. (e.g., invert(data, solver=brd) or invert(data, solver=pdhgm(10,0.1) )).

NMRInversions.brd — Type

brd

Solver for tikhonov (L2) regularization, following this paper from Venkataramanan et al. Very fast, but only uses the identity as tiknohonov matrix. No options required, it just works. It can be used as a "solver" for the invert function.

NMRInversions.pdhgm — Type

pdhgm(σ, τ)

Primal dual hybrid gradient method for L1 regularization, following this paper from Reci et al. / Journal of Magnetic Resonance 281 (2017) 188–198 It can be used as a "solver" for the invert function.

The particular choice of σ and τ is heuristic. A smaller σ will increase the stability while reducing the convergence speed of the algorithm. A good compromise between the two was found when σ = 0.1 and τ = 10. The best values of σ and τ will depend slightly on the scaling of the signal. Therefore, it is best to normalize the NMR signal to a maximum of 1, a technique which was followed in the cited study.

NMRInversions.optim_nnls — Type

optim_nnls(order)

Simple non-negative least squares method for tikhonov (L2) regularization, implemented using OptimizationOptimJl. All around effective, but can be slow for large problems, such as 2D inversions. It can be used as a "solver" for invert function. Order determines the tikhonov matrix. If 0 is chosen, the identity matrix is used.

Finding optimal alpha

These are methods for finding the optimal regularization parameter. They can be used as input to the invert function as the 'alpha' argument (e.g., invert(data, alpha=gcv) ). If you'd like to use a particular value of alpha, you can just use that number instead (invert(data, alpha=1).

NMRInversions.gcv — Type

gcv

Generalized cross validation for finding the optimal regularization parameter α.

NMRInversions.lcurve — Type

lcurve

L curve method for finding the optimal regularization parameter α.

STILL UNDER DEVELOPMENT!

Inversion and expfit inputs

These can be used as inputs containing all of the necessary information to run the inversion and expfit functions (e.g. if a is an input1D object, invert(a) does the job).

NMRInversions.input1D — Type

input1D(seq, x, y)

A structure containing the following fields:

seq is the 1D pulse sequence (e.g. IR, CPMG, PGSE)

x, the x values of the measurement (e.g time for relaxation or b-factor for diffusion).

y, the y values of the measurement.

It can be used as an input for the invert and expfit functions.

NMRInversions.input2D — Type

input2D(seq, x, y)

A structure containing the following fields:

seq is the 2D pulse sequence (e.g. IRCPMG)

x_direct is the direct dimension acquisition parameter (e.g. the times when you aquire CPMG echoes).

x_indirect is the indirect dimension acquisition parameter (e.g. all the delay times τ in your IR sequence).

data is the 2D data matrix.

It can be used as an input for the invert function.

Inversion and expfit outputs

When you run an inversion function (e.g. r = invert(a)), the output is an inv_out_1D or inv_out_2D structure.

NMRInversions.inv_out_1D — Type

inv_out_1D(seq, x, y, xfit, yfit, X, f, r, SNR, α, wa)

Output of the invert function for 1D pulse sequences. A structure containing the following fields:

seq is the 1D pulse sequence (e.g. IR, CPMG, PGSE)

x, the x values of the measurement (e.g time for relaxation or b-factor for diffusion).

y, the y values of the measurement.

xfit, the x values of the fitted data.

yfit, the y values of the fitted data.

X, the x values of the inversion results.

f, the inversion results.

r, the residuals.

SNR, the signal-to-noise ratio.

α, the regularization parameter.

wa, the weighted average of the inversion results (e.g. the mean relaxation time or diffusion coefficient).

NMRInversions.inv_out_2D — Type

inv_out_2D(seq, X_dir, X_indir, F, r, SNR, α, filter, selections)

Output of the invert function for 2D pulse sequences. A structure containing the following fields:

seq is the 2D pulse sequence (e.g. IRCPMG)

X_dir, the x values of the direct dimension.

X_indir, the x values of the indirect dimension.

F, the inversion results as a matrix.

r, the residuals.

SNR, the signal-to-noise ratio.

alpha, the regularization parameter.

filter, apply a mask to filter out artefacts when plotting.

selections, the selection masks (e.g. when you want to highlight some peaks in a T₁-T₂ map).

NMRInversions.expfit_struct — Type

Output of the expfit function. Structure containing information about multiexponential fits.

The fields are as follows:

seq: The pulse sequence

x : The x acquisition values (e.g. time for relaxation or b-factor for diffusion).

y : The y acquisition values.

u : The fitted parameters for the mexp function.

u0 : The initial parameters for the mexp function.

r : The residuals.

eq : The equation of the fitted function.

eqn : The equation of the initial function.

Kernel structure

NMRInversions.svd_kernel_struct — Type

svd_kernel_struct(K,g,U,S,V)

A structure containing the following fields:

K, the kernel matrix.

G, the data vector.

U, the left singular values matrix.

S, the singular values vector.

V, the right singular values matrix.

To access the fields of a structure, we use the dot notation, e.g. if the structure is called a and we want to access the kernel contained in there, we type a.K